fix(stops): robot now actually stops smoothly
The robot was overshooting the stops instead of changing less, providing
the opposite effect than we wanted. I now subtract the difference
between cycles to add back some of the speed that was taken away and
multiply by the constant that now controls how much speed is added back
instead of how much is taken away.